libui: GraphicBuffer: revert to 32-bit usage

+++ libs/ui/include/ui/GraphicBufferAllocator.h
@@ -63,7 +63,7 @@ private:
         uint32_t stride;
         PixelFormat format;
         uint32_t layerCount;
-        uint64_t usage;
+        uint32_t usage;
         size_t size;
         std::string requestorName;
     };
